Time: 448 ms (5.07%) | Memory: 77.3 MB (23.94%) - LeetSync

+++ 1803-average-waiting-time/average-waiting-time.cpp
@@ -3,7 +3,7 @@ public:
     double averageWaitingTime(vector<vector<int>>& customers) {
         int busy=customers[0][0];
         double waiting=0;
-        double n=customers.size();
+        int n=customers.size();
         for(int i=0;i<n;i++){
             busy=max(busy+customers[i][1],customers[i][1]+customers[i][0]);
             cout<<busy<<endl;
